Split error types into separate modules for better organization
- Create error/app_error.rs for AppError and AppResult
- Create error/load_error.rs for LoadError and DuplicateGroup
- Create error/mod.rs with clean re-exports
- Remove monolithic error.rs file

This improves readability by putting AppError (the main application error type)
in its own focused file, while keeping data loading errors separate. The module
structure makes it easier to navigate and understand error handling concerns.
