+++ src/error/mod.rs
@@ -1,0 +1,5 @@
+mod app_error;
+mod load_error;
+
+pub use app_error::{AppError, AppResult};
+pub use load_error::{DuplicateGroup, LoadError};
